// src/types/generated/RecordedCircuitObservation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CircuitObservation } from "./CircuitObservation";
import type { ObservationDisposition } from "./ObservationDisposition";

export type RecordedObservation = { observation: CircuitObservation, disposition: ObservationDisposition, };